requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
elastic-transport==8.15.0
elasticsearch==8.15.0
exceptiongroup==1.2.2
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
idna==3.8
Mako==1.3.5
MarkupSafe==2.1.5
numpy==2.1.1
pandas==2.2.2
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
